Read Desc
removed old logging
factored out fft and factored out n, ijh, and trig stuff for repeated use in fft
normalizing magnitude with bincount, should maybe do that in fft function.
